refactor(scripts): split profile helpers
Move profiling, soak, and A/B report logic behind focused operator helpers while keeping existing script entrypoints and output contracts stable.

Preserve run artifact loading, metric delta calculation, baseline comparison, and report rendering behavior for operator workflows.
