+++ scripts/collect_soak_metrics.py
@@ -3,152 +3,21 @@ from __future__ import annotations
 
 import argparse
 import json
-import subprocess
 import time
 from pathlib import Path
-from typing import Any
-from urllib import request
 
 from scripts.operator_prometheus import PROM_LINE as PROM_LINE
 from scripts.operator_prometheus import parse_metrics as _parse_metrics
 from scripts.operator_prometheus import sum_metric as _sum_metric
-
-WORKER_METRICS_SCRAPE_ATTEMPTS = 2
-WORKER_METRICS_BACKOFF_SECONDS = 0.5
-
-
-def _safe_float(value: Any) -> float | None:
-    try:
-        return float(value)
-    except Exception:
-        return None
-
-
-def _fetch_text(url: str, timeout: int = 10) -> str:
-    with request.urlopen(url, timeout=timeout) as resp:
-        return resp.read().decode("utf-8", errors="replace")
-
-
-def _docker_exec_python(script: str) -> str:
-    cmd = [
-        "docker",
-        "compose",
-        "exec",
-        "-T",
-        "worker",
-        "python",
-        "-c",
-        script,
-    ]
-    return subprocess.check_output(cmd, text=True, stderr=subprocess.STDOUT, timeout=30)
-
-
-def _fetch_worker_metrics_via_docker() -> tuple[str, str | None]:
-    # Strategy order is intentional: prefer the worker metrics HTTP endpoint first,
-    # then fall back to direct process-registry exposition when the endpoint is unavailable.
-    strategies = [
-        (
-            "worker_http",
-            (
-                "import urllib.request; "
-                "print(urllib.request.urlopen('http://localhost:8001/metrics', timeout=10)"
-                ".read().decode('utf-8', errors='replace'))"
-            ),
-        ),
-        (
-            "worker_registry",
-            (
-                "from prometheus_client import CollectorRegistry, generate_latest; "
-                "from pipeline.metrics import RedisProviderMetricsCollector; "
-                "registry = CollectorRegistry(); "
-                "registry.register(RedisProviderMetricsCollector()); "
-                "print(generate_latest(registry).decode('utf-8', errors='replace'))"
-            ),
-        ),
-    ]
-    errors: list[str] = []
-    for strategy_name, script in strategies:
-        for attempt in range(1, WORKER_METRICS_SCRAPE_ATTEMPTS + 1):
-            try:
-                raw = _docker_exec_python(script)
-                if raw.strip():
-                    # HTTP scrape can return generic process metrics without provider
-                    # series; in that case continue to the collector fallback path.
-                    if strategy_name == "worker_http":
-                        provider_sample_present = any(
-                            line.startswith("tc_provider_") and not line.startswith("#")
-                            for line in raw.splitlines()
-                        )
-                        if not provider_sample_present:
-                            errors.append(f"{strategy_name}[attempt={attempt}] missing_provider_series")
-                            continue
-                    return raw, None
-                errors.append(f"{strategy_name}[attempt={attempt}] empty_output")
-            except Exception as exc:
-                errors.append(f"{strategy_name}[attempt={attempt}] {exc}")
-            if attempt < WORKER_METRICS_SCRAPE_ATTEMPTS:
-                time.sleep(WORKER_METRICS_BACKOFF_SECONDS)
-    try:
-        joined = "; ".join(errors).strip()
-        return "", joined or "worker metrics scrape failed"
-    except Exception:
-        return "", "worker metrics scrape failed"
-
-
-def _hist_quantile(rows: list[dict[str, Any]], base_name: str, labels: dict[str, str], quantile: float) -> float | None:
-    buckets: dict[float, float] = {}
-    for row in rows:
-        if row["name"] != f"{base_name}_bucket":
-            continue
-        row_labels = row["labels"]
-        ok = True
-        for key, expected in labels.items():
-            if row_labels.get(key) != expected:
-                ok = False
-                break
-        if not ok:
-            continue
-        le = row_labels.get("le")
-        if le is None:
-            continue
-        if le == "+Inf":
-            upper = float("inf")
-        else:
-            try:
-                upper = float(le)
-            except Exception:
-                continue
-        buckets[upper] = float(row["value"])
-
-    if not buckets:
-        return None
-    sorted_bounds = sorted(buckets.items(), key=lambda x: x[0])
-    total = sorted_bounds[-1][1]
-    if total <= 0:
-        return 0.0
-    target = total * quantile
-    prev_bound = 0.0
-    prev_count = 0.0
-    for bound, count in sorted_bounds:
-        if count >= target:
-            if count == prev_count:
-                return float(bound if bound != float("inf") else prev_bound)
-            frac = (target - prev_count) / (count - prev_count)
-            if bound == float("inf"):
-                return float(prev_bound)
-            return float(prev_bound + (bound - prev_bound) * frac)
-        prev_bound = bound if bound != float("inf") else prev_bound
-        prev_count = count
-    return float(prev_bound)
-
-
-def _provider_metrics_state(rows: list[dict[str, Any]], worker_metrics_error: str | None) -> tuple[bool, str]:
-    if worker_metrics_error:
-        return False, "worker_scrape_failed"
-    provider_series_present = any(str(row.get("name", "")).startswith("tc_provider_") for row in rows)
-    if provider_series_present:
-        return True, "ok"
-    return False, "no_provider_series"
+from scripts.operator_profile_metrics import fetch_text as _fetch_text
+from scripts.operator_profile_metrics import fetch_worker_metrics_via_docker as _fetch_worker_metrics_via_docker
+from scripts.operator_profile_metrics import hist_quantile as _hist_quantile
+from scripts.operator_profile_metrics import load_run_manifest as _load_run_manifest
+from scripts.operator_profile_metrics import load_tasks_rows as _load_tasks_rows
+from scripts.operator_profile_metrics import provider_metrics_state as _provider_metrics_state
+from scripts.operator_profile_metrics import slowest_task as _slowest_task
+from scripts.operator_profile_metrics import submission_failure_breakdown as _submission_failure_breakdown
+from scripts.operator_profile_metric_deltas import provider_run_deltas_from_manifest as _provider_run_deltas_from_manifest
 
 
 def _search_p95_ms(api_url: str) -> float | None:
@@ -166,181 +35,6 @@ def _search_p95_ms(api_url: str) -> float | None:
     ordered = sorted(samples)
     idx = max(0, int((0.95 * len(ordered)) + 0.999999) - 1)
     return float(ordered[idx])
-
-
-def _load_json_file(path: Path) -> dict[str, Any]:
-    try:
-        data = json.loads(path.read_text(encoding="utf-8"))
-    except Exception:
-        return {}
-    return data if isinstance(data, dict) else {}
-
-
-def _load_run_manifest(run_dir: Path) -> dict[str, Any]:
-    return _load_json_file(run_dir / "run_manifest.json")
-
-
-def _load_tasks_rows(run_dir: Path) -> list[dict[str, Any]]:
-    path = run_dir / "tasks.jsonl"
-    if not path.exists():
-        return []
-    rows: list[dict[str, Any]] = []
-    for raw in path.read_text(encoding="utf-8").splitlines():
-        line = raw.strip()
-        if not line:
-            continue
-        try:
-            payload = json.loads(line)
-        except Exception:
-            continue
-        if isinstance(payload, dict):
-            rows.append(payload)
-    return rows
-
-
-def _task_duration(rows: list[dict[str, Any]], phase: str) -> list[tuple[int | None, float]]:
-    out: list[tuple[int | None, float]] = []
-    for row in rows:
-        if row.get("phase") != phase:
-            continue
-        try:
-            duration = float(row.get("duration_s") or 0.0)
-        except Exception:
-            duration = 0.0
-        catalog_id = row.get("catalog_id")
-        try:
-            catalog_value = int(catalog_id)
-        except Exception:
-            catalog_value = None
-        out.append((catalog_value, duration))
-    return out
-
-
-def _slowest_task(rows: list[dict[str, Any]]) -> dict[str, Any]:
-    slowest_phase = ""
-    slowest_catalog_id = None
-    slowest_duration_s = 0.0
-    segment_max_s = 0.0
-    summary_max_s = 0.0
-
-    for row in rows:
-        try:
-            duration = float(row.get("duration_s") or 0.0)
-        except Exception:
-            duration = 0.0
-        phase = str(row.get("phase") or "")
-        catalog_id = row.get("catalog_id")
-        try:
-            catalog_value = int(catalog_id)
-        except Exception:
-            catalog_value = None
-
-        if duration >= slowest_duration_s:
-            slowest_phase = phase
-            slowest_catalog_id = catalog_value
-            slowest_duration_s = duration
-        if phase == "segment":
-            segment_max_s = max(segment_max_s, duration)
-        if phase == "summarize":
-            summary_max_s = max(summary_max_s, duration)
-
-    return {
-        "slowest_phase": slowest_phase or None,
-        "slowest_catalog_id": slowest_catalog_id,
-        "slowest_duration_s": float(slowest_duration_s),
-        "segment_max_s": float(segment_max_s),
-        "summary_max_s": float(summary_max_s),
-    }
-
-
-def _submission_failure_breakdown(rows: list[dict[str, Any]]) -> dict[str, int]:
-    counts = {
-        "task_submission_failures": 0,
-        "task_submission_error_failures": 0,
-        "unexpected_non_processing_status_failures": 0,
-        "missing_task_id_failures": 0,
-        "task_poll_timeouts": 0,
-    }
-    for row in rows:
-        error = str(row.get("error") or "").strip()
-        if not error:
-            continue
-        if error == "task_submission_error":
-            counts["task_submission_failures"] += 1
-            counts["task_submission_error_failures"] += 1
-        elif error.startswith("unexpected_non_processing_status:"):
-            counts["task_submission_failures"] += 1
-            counts["unexpected_non_processing_status_failures"] += 1
-        elif error == "missing_task_id":
-            counts["task_submission_failures"] += 1
-            counts["missing_task_id_failures"] += 1
-        elif error == "task_poll_timeout":
-            counts["task_poll_timeouts"] += 1
-    return counts
-
-
-def _provider_run_deltas_from_manifest(
-    manifest: dict[str, Any],
-    *,
-    provider_requests_total: float,
-    provider_timeouts_total: float,
-    provider_retries_total: float,
-) -> dict[str, float | None]:
-    baseline = manifest.get("provider_counters_before_run")
-    if not isinstance(baseline, dict):
-        return {
-            "provider_requests_delta_run": None,
-            "provider_timeouts_delta_run": None,
-            "provider_retries_delta_run": None,
-            "provider_timeout_rate_run": None,
-        }
-
-    baseline_keys = (
-        "provider_requests_total",
-        "provider_timeouts_total",
-        "provider_retries_total",
-    )
-    legacy_keys = (
-        "tc_provider_requests_total",
-        "tc_provider_timeouts_total",
-        "tc_provider_retries_total",
-    )
-    if any(key in baseline for key in legacy_keys):
-        return {
-            "provider_requests_delta_run": None,
-            "provider_timeouts_delta_run": None,
-            "provider_retries_delta_run": None,
-            "provider_timeout_rate_run": None,
-        }
-    if any(key not in baseline for key in baseline_keys):
-        return {
-            "provider_requests_delta_run": None,
-            "provider_timeouts_delta_run": None,
-            "provider_retries_delta_run": None,
-            "provider_timeout_rate_run": None,
-        }
-
-    baseline_requests = _safe_float(baseline.get("provider_requests_total"))
-    baseline_timeouts = _safe_float(baseline.get("provider_timeouts_total"))
-    baseline_retries = _safe_float(baseline.get("provider_retries_total"))
-    if baseline_requests is None or baseline_timeouts is None or baseline_retries is None:
-        return {
-            "provider_requests_delta_run": None,
-            "provider_timeouts_delta_run": None,
-            "provider_retries_delta_run": None,
-            "provider_timeout_rate_run": None,
-        }
-
-    requests = max(0.0, float(provider_requests_total - baseline_requests))
-    timeouts = max(0.0, float(provider_timeouts_total - baseline_timeouts))
-    retries = max(0.0, float(provider_retries_total - baseline_retries))
-    timeout_rate = float(timeouts / requests) if requests > 0 else None
-    return {
-        "provider_requests_delta_run": requests,
-        "provider_timeouts_delta_run": timeouts,
-        "provider_retries_delta_run": retries,
-        "provider_timeout_rate_run": timeout_rate,
-    }
 
 
 def main() -> int:
